cortex/api/requirements.txt
# cortex/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis[hiredis]==5.2.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
asyncssh==2.18.0
pydantic==2.10.5
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.0
docker==7.1.0
pytest==8.3.0
pytest-asyncio==0.25.0
